Created the library and connected nodejs with MySQL.
I tried to create the library and insert the content according to what I learned, but I encountered difficulties in connection. I spent two days looking for these problems, and finally configured the system environment variables and optimized the code of js file, and downloaded the dependencies. The final result is here.