Completion of part of Part3
I studied post from 7: 00 p.m. to 4: 50 a.m. But just now, I found that this assignment doesn't need to use post, and there is no requirement for registration and login. This is my negligence. I neglected some content requirements of the assignment, and I will continue to work hard. This is the directory after I deleted the files such as server.js At the same time, I still finished some front-end production while studying post today. I uploaded this, and I am considering whether to delete the front-end login page or keep it as a warning that I need to read the homework requirements well.